Added Comments

+++ project/routes.py
@@ -2,9 +2,9 @@ from flask import *
 import os
 from project import app, db, bcrypt
 from datetime import datetime
-from project.model import Post, User
-from project.form import RegistrationForm, LoginForm
-from flask_login import current_user, login_user, logout_user
+from project.model import Post, User, Comment
+from project.form import RegistrationForm, LoginForm, CommentForm
+from flask_login import current_user, login_user, logout_user, login_required
 from sqlalchemy import desc, asc
 from werkzeug.utils import secure_filename
 
@@ -27,47 +27,66 @@ def index():
 
 # Admin Panel pages
 @app.route('/admin', methods=['GET', 'POST'])
+@login_required
 def all_posts():
-    form = RegistrationForm()
-    posts = Post.query.order_by(desc(Post.date_posted)).all()
-    return render_template("admin_panel/all_posts.html", posts=posts, form=form)
+    if current_user.permissions == 'admin':
+        form = RegistrationForm()
+        posts = Post.query.order_by(desc(Post.date_posted)).all()
+        return render_template("admin_panel/all_posts.html", posts=posts, form=form)
+    else:
+        return redirect('/')
 
 
 @app.route('/admin/modify', methods=['GET', 'POST'])
+@login_required
 def modify():
-    form = LoginForm()
-    posts = Post.query.order_by(desc(Post.date_posted)).all()
-    return render_template("admin_panel/modify_posts.html", posts=posts, form=form)
+    if current_user.permissions == 'admin':
+        form = LoginForm()
+        posts = Post.query.order_by(desc(Post.date_posted)).all()
+        return render_template("admin_panel/modify_posts.html", posts=posts, form=form)
+    else:
+        return redirect('/')
 
 
 @app.route('/admin/users', methods=['GET', 'POST'])
+@login_required
 def all_users():
-    users = User.query.all()
-    return render_template("admin_panel/all_users.html", users=users)
+    if current_user.permissions == 'admin':
+        users = User.query.all()
+        return render_template("admin_panel/all_users.html", users=users)
+    else:
+        return redirect('/')
 
 # Adding user
 @app.route('/admin/add_user', methods=['GET', 'POST'])
+@login_required
 def add_user():
-    form = RegistrationForm()
-    if form.validate_on_submit():
-        hashed_password = bcrypt.generate_password_hash(
-            form.password.data).decode('utf-8')
-        user = User(username=form.username.data, email=form.email.data,
-                    password=hashed_password, permissions=form.permission.data)
-        db.session.add(user)
-        db.session.commit()
-        return redirect('/')
+    if current_user.permissions == 'admin':
+        form = RegistrationForm()
+        if form.validate_on_submit():
+            hashed_password = bcrypt.generate_password_hash(
+                form.password.data).decode('utf-8')
+            user = User(username=form.username.data, email=form.email.data,
+                        password=hashed_password, permissions=form.permission.data)
+            db.session.add(user)
+            db.session.commit()
+            return redirect('/')
+        else:
+            return render_template("admin_panel/new_user.html", form=form)
     else:
-        return render_template("admin_panel/new_user.html", form=form)
-
+        return redirect('/')
 
 # Delating user
 @app.route('/users/delete_user/<int:id>', methods=['GET', 'POST'])
+@login_required
 def delete_user(id):
-    user = User.query.get_or_404(id)
-    db.session.delete(user)
-    db.session.commit()
-    return redirect('/admin/users')
+    if current_user.permissions == 'admin':
+        user = User.query.get_or_404(id)
+        db.session.delete(user)
+        db.session.commit()
+        return redirect('/admin/users')
+    else:
+        return redirect('/')
 
 # Login in
 @app.route('/login', methods=['GET', 'POST'])
@@ -127,58 +146,97 @@ def allowed_image(filename):
 
 # Posts interactions (editing, deleting, creating new)
 @app.route('/admin/new', methods=['GET', 'POST'])
+@login_required
 def new():
-    if request.method == "POST":
-        if request.files:
-            image = request.files["file"]
-            image.save(os.path.join(
-                app.config["IMAGE_UPLOADS"], image.filename))
-            print(image)
+    if current_user.permissions == 'admin':
+        if request.method == "POST":
+            if request.files:
+                image = request.files["file"]
+                image.save(os.path.join(
+                    app.config["IMAGE_UPLOADS"], image.filename))
+                print(image)
 
-        post_image_file = image.filename
-        post_title = request.form['title']
-        post_content = request.form['content']
-        post_author = request.form['author']
-        new_post = Post(image_file=post_image_file, title=post_title,
-                        content=post_content, author=post_author)
-        db.session.add(new_post)
-        db.session.commit()
-        print("gitara")
-        return redirect('/')
+            post_image_file = image.filename
+            post_title = request.form['title']
+            post_content = request.form['content']
+            post_author = request.form['author']
+            new_post = Post(image_file=post_image_file, title=post_title,
+                            content=post_content, author=post_author)
+            db.session.add(new_post)
+            db.session.commit()
+            return redirect('/')
+        else:
+            return render_template("admin_panel/new_post.html")
     else:
-        return render_template("admin_panel/new_post.html")
+        return redirect('/')
 
 
-@app.route('/posts/more/<int:id>', methods=['GET', 'POST'])
+@app.route('/post/more/<int:id>', methods=['GET', 'POST'])
 def more(id):
+    select = request.form.get('comment_select')
+    if str(select) == '1':
+        comments = Comment.query.order_by(asc(Comment.time)).all()
+    else:
+        comments = Comment.query.order_by(desc(Comment.time)).all()
+
     form = RegistrationForm()
     posts = Post.query.get_or_404(id)
-    return render_template("post.html", posts=posts, form=form)
+    comments_count = Comment.query.filter_by(post_id=id).count()
+    users = User.query.all()
+
+    Commentform = CommentForm()
+    if Commentform.validate_on_submit():
+        author_id = current_user.id
+        comment = Comment(author=author_id,
+                          body=Commentform.Comment.data, post_id=id)
+        db.session.add(comment)
+        db.session.commit()
+        return redirect(request.url)
+    return render_template("post.html", posts=posts, form=form, comment=Commentform, 
+    comments=comments, users=users, comments_count=comments_count)
+
+
+@app.route('/post/more/deletecom/<int:id>', methods=['GET', 'POST'])
+@login_required
+def deletecom(id):
+    if current_user.permissions == 'admin':
+        comment = Comment.query.get_or_404(id)
+        db.session.delete(comment)
+        db.session.commit()
+        return redirect('/')
 
 
 @app.route('/posts/delete/<int:id>', methods=['GET', 'POST'])
+@login_required
 def delete(id):
-    post = Post.query.get_or_404(id)
-    db.session.delete(post)
-    db.session.commit()
-    return redirect('/')
+    if current_user.permissions == 'admin':
+        post = Post.query.get_or_404(id)
+        db.session.delete(post)
+        db.session.commit()
+        return redirect('/posts/modify')
+    else:
+        return redirect('/')
 
 
 @app.route('/posts/edit/<int:id>', methods=['GET', 'POST'])
+@login_required
 def edit(id):
     post = Post.query.get_or_404(id)
-    if request.method == "POST":
-        if request.files:
-            image = request.files["edit_file"]
-            image.save(os.path.join(
-                app.config["IMAGE_UPLOADS"], image.filename))
-            print(image)
+    if current_user.permissions == 'admin':
+        if request.method == "POST":
+            if request.files:
+                image = request.files["edit_file"]
+                image.save(os.path.join(
+                    app.config["IMAGE_UPLOADS"], image.filename))
+                print(image)
 
-        post._image_file = image.filename
-        post.title = request.form['edit_title']
-        post.content = request.form['edit_content']
-        post.author = request.form['edit_author']
-        db.session.commit()
-        return redirect('/posts')
+            post._image_file = image.filename
+            post.title = request.form['edit_title']
+            post.content = request.form['edit_content']
+            post.author = request.form['edit_author']
+            db.session.commit()
+            return redirect('/posts')
+        else:
+            return render_template("admin_panel/edit_posts.html", post=post)
     else:
-        return render_template("admin_panel/edit_posts.html", post=post)
+        return redirect('/')
